+++ src/main/java/com/dkkm/marketsim/service/ClosingService.java
@@ -2,9 +2,10 @@ package com.dkkm.marketsim.service;
 
 import com.dkkm.marketsim.model.dto.Closing;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public interface ClosingService extends PassThruCrudService<Closing, Closing> {
 
-    double getSharePrice(String ticker, LocalDate date);
+    BigDecimal getSharePrice(String ticker, LocalDate date);
 }
